+++ packages/terra-clinical-item-display/tests/nightwatch/display/TextStyleItemDisplay.jsx
@@ -7,6 +7,7 @@ const views = () => (
     <ItemDisplay text="test text" textStyle="secondary" id="test-display-secondary" />
     <ItemDisplay text="test text" textStyle="attention" id="test-display-attention" />
     <ItemDisplay text="test text" textStyle="strikeThrough" id="test-display-strike-through" />
+    <ItemDisplay text="test text" textStyle="strong" id="test-display-strong" />
   </div>
 );