+++ packages/terra-clinical-site/src/examples/item-display/Icon.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
-
 import ItemDisplay from 'terra-clinical-item-display';
 import IconGlasses from 'terra-icon/lib/icon/IconGlasses';
 
-const TextIconItemDisplay = () => (<ItemDisplay text="Display" icon={<IconGlasses />} />);
+const component = () => (<ItemDisplay icon={<IconGlasses />} />);
 
-export default TextIconItemDisplay;
+export default component;